\name{scatter_gate}
\alias{scatter_gate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Singlets Detection}
\description{
Gate by an ellipse of radius R on the scatter data.
}
\usage{
scatter_gate(fcs, scatters, R = 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fcs}{
FCS file.
}
  \item{scatters}{
Indices of the Forward scatter and Side Scatter in the FCS file.
}
  \item{R}{
Gate by an ellipse of radius R on the scatter data.

}
}
\details{
Singlets detection using Forward scatter and Side Scatter in the FCS file.
}
\value{
A new FCS file with only singlets.
}
\references{
Wayne Moore et al. 2014, Automated Quadratic Characterization of Flow Cytometer Instrument Sensitivity, to be submitted. 
}
\author{
Wayne Moore
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
